pyproject.toml
[tool.poetry]
name = "imgs-to-tiff"
version = "0.1.0"
description = ""
authors = ["NickolayBabulich <nickolas.grn@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
